Initial commit of SW revision 2.0.0_ALPHA
Software revision 2.0.0 no longer operates decentralized, but instead
now makes use of a local network connection.

Bells:
The bells are constantly connected to our wireless network and await a
TCP package from the door. Once a valid TCP package has been received,
the bell will ring.

Door:
Upon button press, the transmitter will establish a connection to the
wireless network and send out a RING_MSG tcp packet to all bells
attached to the network.

The software relies on a static pool of host IDs. At the moment, the door
must be assigned a host ID ending with 0 (ex. 192.168.0.20). The bells should
be assigned to the following up host ID's (ex. 192.168.0.21/22/23...).

As the title of this commit already says, this is an alpha build, meaning that it is
potentially buggy, and certainly requires some refactoring. However, despite being
an alpha build, the decrease in propagation delay and increase in transmission
reliability already makes this version significanlty more stable than the previous
decentralizied one.

Signed-off-by: CTXz <ctx.xda@gmail.com>