v2.1.0: Major refactor of door code
The door code has been completely overhauled and now consists of a main
Door class, which is configured by a DoorCFG object, as well has a WiFiHandler class, a RingSender class and
a RingTX class.

To ensure processes run "asynchronously", all classes are implemented as state machines.

Door
----
The main Door class.
This class handles the main logic of the door transmitter. Everything from initialization,
to error handling and to most importantly instancing and utilizing the WifiHandler and RingSender.

DoorCFG
---
Builder class for Door class.
The DoorCFG class is used to configure a Door class instance. It exposes a variety of parameter members
as well as a function to check if it is valid (this is used by the Door class to check if the config is valid).

WiFiHandler
---
Establishes and monitors a WiFi connection.
The WiFi handler is used to establish and monitor an active WiFi connection.

RingSender
---
Sends out ring messages to all bells.
The RingSender sends out ring messages to all targeted bells. It does so
using RingTX objects, which are responsible for sending a ring message
to a targeted bell.

RingTX
---
Transmits a ring message to a bell.
The RingTX transmits a ring message to a bell. It does so using @me-no-dev's AsyncTCP lib.

Signed-off-by: CTXz <ctx.xda@gmail.com>