Bell: v2.1.0: Major refactor of bell code
This commit is the second part of our major refactor.

This time the bell code has been overhauled in similar manner to the
door code.

A main Bell class has been created, which is initialized by a BellCFG
class. WiFi handling is now managed through the shared WiFiHandler.
Ring receiving is handled by the RingReceiver class and the buzzer
is now handled by a Buzzer class.

To ensure processes run "asynchronously", all classes are implemented as state machines.

Bell
---
The main bell class.
This class handles the main logic of the Bell receiver.

BellCFG
---
Builder class for Bell class.
The BellCFG class is used to configure a Bell class instance. It exposes a variety of parameter members
as well as a function to check if it is valid (this is used by the Door class to check if the config is valid).

WiFiHandler
---
Establishes and monitors a WiFi connection.
The WiFi handler is used to establish and monitor an active WiFi connection.

RingReceiver
---
Creates a server which awaits and handles clients and any incoming tcp packets. It verifies
if the incoming client is a door, and whether the received packet is a ring message.
Due to the fact that the asynchronous handling of incoming traffic is only possible through
callbacks (at least with the AsyncTCP lib), referencing static variables is a requirement.
To all static data exchange encapsuled in the class, it has been defined as a singleton.
The singleton can be instanced or retrieved with the get_instance() method.
Since the class does not have a public constructor, it must be initialized with the
begin method.

Buzzer
---
Handles the buzzer (ex. playing a ring melody)

Signed-off-by: CTXz <ctx.xda@gmail.com>